+++ teku/src/integration-test/java/tech/pegasys/teku/cli/subcommand/ValidatorClientCommandTest.java
@@ -33,8 +33,8 @@ import org.mockito.ArgumentCaptor;
 import org.mockserver.integration.ClientAndServer;
 import org.mockserver.junit.jupiter.MockServerExtension;
 import org.mockserver.matchers.Times;
-import org.mockserver.model.HttpError;
 import org.mockserver.model.JsonBody;
+import org.mockserver.socket.PortFactory;
 import tech.pegasys.teku.api.ConfigProvider;
 import tech.pegasys.teku.cli.BeaconNodeCommand;
 import tech.pegasys.teku.config.TekuConfiguration;
@@ -54,9 +54,11 @@ public class ValidatorClientCommandTest {
 
   private String[] argsNetworkOptDefault;
   private String[] argsNetworkOptAuto;
+  private String[] argsNetworkOptAutoWithFallback;
   private String[] argsNetworkOptAutoInConfig;
 
   private ClientAndServer mockBeaconServer;
+  private ClientAndServer fallbackMockBeaconServer;
 
   private final Spec testSpec =
       SpecFactory.create(
@@ -73,13 +75,25 @@ public class ValidatorClientCommandTest {
   @BeforeEach
   public void setup(ClientAndServer server) {
     this.mockBeaconServer = server;
-    String mockBeaconServerEndpoint =
-        String.format("http://127.0.0.1:%s/", mockBeaconServer.getLocalPort());
+    this.fallbackMockBeaconServer =
+        ClientAndServer.startClientAndServer(PortFactory.findFreePort());
+    final String mockBeaconServerEndpoint = getMockBeaconServerEndpoint(mockBeaconServer);
+    final String fallbackMockBeaconServerEndpoint =
+        getMockBeaconServerEndpoint(fallbackMockBeaconServer);
+
     argsNetworkOptDefault =
         new String[] {"vc", "--beacon-node-api-endpoint", mockBeaconServerEndpoint};
     argsNetworkOptAuto =
         new String[] {
           "vc", "--network", "auto", "--beacon-node-api-endpoint", mockBeaconServerEndpoint
+        };
+    argsNetworkOptAutoWithFallback =
+        new String[] {
+          "vc",
+          "--network",
+          "auto",
+          "--beacon-node-api-endpoints",
+          mockBeaconServerEndpoint + "," + fallbackMockBeaconServerEndpoint
         };
     argsNetworkOptAutoInConfig =
         new String[] {
@@ -94,6 +108,7 @@ public class ValidatorClientCommandTest {
   @AfterEach
   public void tearDown() {
     mockBeaconServer.reset();
+    fallbackMockBeaconServer.stop();
   }
 
   @Test
@@ -106,6 +121,22 @@ public class ValidatorClientCommandTest {
   public void autoDetectNetwork_ShouldFetchNetworkDetailsFromBeaconNode_IfEnabled() {
     configureMockServer(0);
     fetchAndVerifySpec(argsNetworkOptAuto);
+  }
+
+  @Test
+  public void autoDetectNetwork_ShouldFetchNetworkDetailsFromFallbackNode() {
+    // primary node always fails
+    configureMockServer(-1);
+    configureSuccessfulResponse(fallbackMockBeaconServer);
+    fetchAndVerifySpec(argsNetworkOptAutoWithFallback);
+  }
+
+  @Test
+  public void autoDetectNetwork_ShouldRetryRequest_IfFailsToFetchFromAllConfiguredBeaconNodes() {
+    configureMockServer(1);
+    // fallback node always fails
+    configureFailedResponse(fallbackMockBeaconServer, -1);
+    fetchAndVerifySpec(argsNetworkOptAutoWithFallback);
   }
 
   @Test
@@ -129,13 +160,26 @@ public class ValidatorClientCommandTest {
   }
 
   private void configureMockServer(int fails) {
-    this.mockBeaconServer
-        .when(request().withPath("/eth/v1/config/spec"), Times.exactly(fails))
-        .error(HttpError.error());
+    configureFailedResponse(mockBeaconServer, fails);
+    configureSuccessfulResponse(mockBeaconServer);
+  }
 
-    this.mockBeaconServer
+  private void configureSuccessfulResponse(final ClientAndServer mockBeaconServer) {
+    mockBeaconServer
         .when(request().withPath("/eth/v1/config/spec"))
         .respond(response().withStatusCode(200).withBody(getTestSpecResponse()));
+  }
+
+  private void configureFailedResponse(final ClientAndServer mockBeaconServer, final int fails) {
+    mockBeaconServer
+        .when(
+            request().withPath("/eth/v1/config/spec"),
+            fails == -1 ? Times.unlimited() : Times.exactly(fails))
+        .respond(response().withStatusCode(500));
+  }
+
+  private String getMockBeaconServerEndpoint(final ClientAndServer mockBeaconServer) {
+    return String.format("http://127.0.0.1:%s/", mockBeaconServer.getLocalPort());
   }
 
   private String getTestSpecResponse() {